pyproject.toml
[project]
name = "exam"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Amjad Alsharafi", email = "26300843+Amjad50@users.noreply.github.com" }
]
dependencies = [
    "django>=5.1",
    "psycopg2>=2.9.9",
    "drf-spectacular[sidecar]>=0.27.2",
    "djangorestframework>=3.15.2",
    "django-auditlog>=3.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
